requirements.txt
voluptuous>=0.11.1
platformio>=3.5.3
pyyaml>=3.12
paho-mqtt>=1.3.1
colorlog>=3.1.2
tornado>=5.0.0
esptool>=2.3.1
typing>=3.0.0
protobuf>=3.4
pyserial>=3.4,<4
ruamel.yaml
